Conflicts:
	src/utils.hpp

+++ admin/static/coffee/resolve_issues.coffee
@@ -38,18 +38,22 @@ module 'ResolveIssuesView', ->
             @machine_to_kill = _machine_to_kill
             super
                 machine_name: @machine_to_kill.get("name")
-                modal_title: "Declare server dead"
-                btn_primary_text: 'Declare Dead'
+                modal_title: "Permanently remove the server"
+                btn_primary_text: 'Remove'
 
         on_submit: ->
             super
 
-            $.ajax
-                url: "ajax/semilattice/machines/#{@machine_to_kill.id}"
-                type: 'DELETE'
-                contentType: 'application/json'
-                success: @on_success
-                error: @on_error
+            if @$('.verification').val().toLowerCase() is 'remove'
+                $.ajax
+                    url: "ajax/semilattice/machines/#{@machine_to_kill.id}"
+                    type: 'DELETE'
+                    contentType: 'application/json'
+                    success: @on_success
+                    error: @on_error
+            else
+                @.$('.error_verification').slideDown 'fast'
+                @reset_buttons()
 
         on_success_with_error: =>
             @.$('.error_answer').html @template_issue_error
@@ -408,14 +412,13 @@ module 'ResolveIssuesView', ->
                 namespace_name: namespaces.get(@model.get('namespace_id')).get('name')
                 datacenters_with_issues: []
 
-
             namespace = namespaces.get(@model.get('namespace_id'))
             if @model.get('primary_datacenter') isnt universe_datacenter.get('id') and not datacenters.get(@model.get('primary_datacenter'))?
                 json.no_primary = true
             else
                 # known issues = issue where replica > number of machines in a datacenter (not universe)
                 number_machines_universe_can_use_if_no_known_issues = machines.length
-                if @model.get('primary_datacenter') is universe_datacenter
+                if @model.get('primary_datacenter') is universe_datacenter.get('id')
                     number_machines_universe_can_use_if_no_known_issues-- #-1 because there is a primary somewhere
 
                 # Find the datacenters in which we are sure that there is a unsatisfiable goal ( replicas > number of machines in the datacenter )
@@ -443,14 +446,17 @@ module 'ResolveIssuesView', ->
                         else
                             number_machines_in_datacenter = 0
                         if datacenter_id isnt universe_datacenter.get('id') and number_replicas > number_machines_in_datacenter
-                            datacenter_name = datacenters.get(datacenter_id).get('name') # That's safe, datacenters.get(datacenter_id) is defined
+                            if datacenter_id is @model.get('primary_datacenter') and number_machines_in_datacenter is 0
+                                json.primary_empty = true
+                            else
+                                datacenter_name = datacenters.get(datacenter_id).get('name') # That's safe, datacenters.get(datacenter_id) is defined
 
-                            json.datacenters_with_issues.push
-                                datacenter_id: datacenter_id
-                                datacenter_name: datacenter_name
-                                num_replicas: number_replicas
-                                num_machines: number_machines_in_datacenter
-                                change_ack: namespace.get('ack_expectations')[datacenter_id].expectation > number_machines_in_datacenter
+                                json.datacenters_with_issues.push
+                                    datacenter_id: datacenter_id
+                                    datacenter_name: datacenter_name
+                                    num_replicas: number_replicas
+                                    num_machines: number_machines_in_datacenter
+                                    change_ack: namespace.get('ack_expectations')[datacenter_id].expectation > number_machines_in_datacenter
 
                         # We substract the number of machines used by the datacenter if we solve the issue
                         if datacenter_id isnt universe_datacenter.get('id')
@@ -473,12 +479,15 @@ module 'ResolveIssuesView', ->
                         else
                             number_machines_in_datacenter = 0
 
-                        json.datacenters_with_issues.push
-                            datacenter_id: datacenter_id
-                            datacenter_name: datacenters.get(datacenter_id).get('name') # Safe since it cannot be universe
-                            num_replicas: number_replicas
-                            num_machines: 0
-                            change_ack: namespace.get('ack_expectations')[datacenter_id].expectation > number_machines_in_datacenter
+                        if datacenter_id is @model.get('primary_datacenter') and number_machines_in_datacenter is 0
+                            json.primary_empty = true
+                        else
+                            json.datacenters_with_issues.push
+                                datacenter_id: datacenter_id
+                                datacenter_name: datacenters.get(datacenter_id).get('name') # Safe since it cannot be universe
+                                num_replicas: number_replicas
+                                num_machines: 0
+                                change_ack: namespace.get('ack_expectations')[datacenter_id].expectation > number_machines_in_datacenter
 
 
                 number_machines_requested_by_universe = @model.get('replica_affinities')[universe_datacenter.get('id')]
@@ -525,14 +534,18 @@ module 'ResolveIssuesView', ->
                                     datacenter_name: (datacenter_name if datacenter_name?)
                                     num_replicas_requested: num_replicas_requested
 
-                        # Let's add universe at the beginning
+                        # Let's add universe at the beginning of the list of datacenters that require some replicas
+                        num_replicas_universe_request = @model.get('replica_affinities')[universe_datacenter.get('id')]
+                        if @model.get('primary_datacenter') is universe_datacenter.get('id')
+                            num_replicas_universe_request += 1
                         json.extra_replicas_accross_cluster.datacenters_that_can_help.unshift
                             datacenter_id: universe_datacenter.get('id')
                             is_universe: true
-                            num_replicas_requested: @model.get('replica_affinities')[universe_datacenter.get('id')]
+                            num_replicas_requested: num_replicas_universe_request
 
 
                 json.can_solve_issue = json.datacenters_with_issues.length > 0
+
             @.$el.html _template(json)
 
             # bind resolution events
